+++ backend/core/log_streamer/__init__.py
@@ -1,0 +1,8 @@
+"""
+   Copyright (C) Mov.ai  - All Rights Reserved
+   Unauthorized copying of this file, via any medium is strictly prohibited
+   Proprietary and confidential
+
+   Developers:
+   - Erez Zomer  (erez@mov.ai) - 2022
+"""
